level_one/newsapi/requirements.txt
apiclient==1.0.4
appnope==0.1.0
asgiref==3.2.7
backcall==0.1.0
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
Django==3.0.4
django-extensions==2.2.9
djangorestframework==3.11.0
entrypoints==0.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.1
idna==2.9
importlib-metadata==1.6.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
mccabe==0.6.1
oauth2client==4.1.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
ujson==1.35
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
